+++ app/src/main/res/layout/activity_main.xml
@@ -14,14 +14,7 @@
         android:layout_height="match_parent"
         android:layout_above="@+id/bvBottomNavigation">
 
-        <com.google.android.material.floatingactionbutton.FloatingActionButton
-            android:id="@+id/fabCompose"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="bottom|right"
-            android:layout_margin="16dp"
-            android:src="@drawable/home_edit_outline"
-            app:fabSize="normal" />
+
     </FrameLayout>
 
     <com.google.android.material.bottomnavigation.BottomNavigationView
