+++ app/src/main/res/layout/fragment_edit_profile.xml
@@ -31,7 +31,7 @@
                     android:id="@+id/topAppBar"
                     android:layout_width="match_parent"
                     android:layout_height="?attr/actionBarSize"
-                    app:navigationIcon="@drawable/ic_back"
+
                     >
 
                     <TextView
@@ -39,9 +39,10 @@
                         android:layout_width="wrap_content"
                         android:layout_height="wrap_content"
                         android:layout_centerHorizontal="false"
-                        android:foregroundGravity="center_horizontal"
                         android:layout_gravity="center_horizontal"
+                        android:foregroundGravity="center_horizontal"
                         android:text="@{user.getKeyUsername()}"
+                        android:textAppearance="@style/TextAppearance.AppCompat.Medium"
                         android:textColor="@color/orange"
                         android:textStyle="bold" />
 
@@ -260,7 +261,8 @@
                     android:layout_below="@+id/passwordLayout"
                     android:layout_marginLeft="45dp"
                     android:layout_marginTop="30dp"
-                    android:text="SAVE"/>
+                    android:layout_marginBottom="30dp"
+                    android:text="Save"/>
 
 
             </LinearLayout>
