+++ app/src/main/res/layout/item_post.xml
@@ -113,6 +113,8 @@
             android:textSize="16sp"
             android:textColor="@color/black"
             android:text="@{post.getQuestion()}"
+            android:autoLink="all"
+            android:linksClickable="true"
             tools:text="Testing Testing Testing Testing Testing Testing Testing Testing Testing Testing Testing Testing Testing Testing Testing Testing Testing Testing " />
 
 
@@ -150,9 +152,9 @@
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_marginEnd="12dp"
-                android:text="LIKES"
+                android:text="Likes"
                 android:visibility="@{post.getLikesCount()!=null ? View.VISIBLE : View.GONE}"
-                tools:text="LIKES" />
+                 />
 
             <TextView
                 android:id="@+id/tvCommentsCount"
@@ -167,9 +169,9 @@
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_marginEnd="12dp"
-                android:text="COMMENTS"
+                android:text="Comments"
                 android:visibility="@{post.getCommentsCount()!=null ? View.VISIBLE : View.GONE}"
-                tools:text="COMMENTS" />
+                 />
 
         </LinearLayout>
         <View
@@ -193,6 +195,8 @@
                 android:layout_height="40dp"
                 style="@style/Widget.MaterialComponents.Button.TextButton.Icon"
                 app:icon="@drawable/heart_outline"
+                app:iconTint="@color/grey"
+                android:textColor="@color/grey"
                 android:text="LIKE" />
             <Button
                 android:id="@+id/btnComment"
@@ -201,6 +205,8 @@
                 android:layout_alignParentEnd="true"
                 style="@style/Widget.MaterialComponents.Button.TextButton.Icon"
                 app:icon="@drawable/message_outline"
+                app:iconTint="@color/grey"
+                android:textColor="@color/grey"
                 android:text=" COMMENT" />
         </RelativeLayout>
