+++ app/src/main/res/layout/post_detail.xml
@@ -46,6 +46,7 @@
                         android:layout_gravity="center_horizontal"
                         android:layout_marginBottom="340dp"
                         android:foregroundGravity="center_horizontal"
+                        android:textAppearance="@style/TextAppearance.AppCompat.Medium"
                         android:text="@{post.getAuthor().getUsername()}"
                         android:textColor="@color/orange"
                         android:textStyle="bold" />
